+++ Changelog.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.9.1] - 2023-07-31
+
+### Fixed
+
+- Error caused by ELtruismNPSettings (Function wasn't in the profile.)
+- Error caused by General (Removed the line it was a mistake)
+- Profile by adding functions, nameplate settings + some tweaks
+- Some typos and general code neatness
+
 ## [0.9.0] - 2023-07-30
 
 ### Added